requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
astroid=2.2.5=py37_0
autopep8=1.4.3=py37_0
blas=1.0=mkl
ca-certificates=2019.1.23=0
certifi=2019.3.9=py37_0
cffi=1.12.2=py37h2e261b9_1
cudatoolkit=10.0.130=0
cudnn=7.3.1=cuda10.0_0
intel-openmp=2019.1=144
isort=4.3.15=py37_0
lazy-object-proxy=1.3.1=py37h14c3975_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=8.2.0=hdf63c60_1
mccabe=0.6.1=py37_1
mkl=2019.1=144
mkl_fft=1.0.10=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py37h6bb024c_1
numpy=1.16.2=py37h7e9f1db_0
numpy-base=1.16.2=py37hde5b4d6_0
openssl=1.1.1b=h7b6447c_1
pandas=0.24.2=py37he6710b0_0
pip=19.0.3=py37_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pylint=2.3.1=py37_0
python=3.7.2=h0371630_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=cuda100py37he554f03_0
pytz=2018.9=py37_0
readline=7.0=h7b6447c_5
setuptools=40.8.0=py37_0
six=1.12.0=py37_0
sqlite=3.27.2=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.33.1=py37_0
wrapt=1.11.1=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
